created a separated view for miltosne board bus
